+++ previous-versions/commerce/pkg/models/shared/address.go
@@ -16,5 +16,54 @@ type Address struct {
 	// The fourth line of the address, or region
 	Region *string `json:"region,omitempty"`
 	// The type of the address
-	Type *AddressType `json:"type,omitempty"`
+	Type *CommerceAddressType `json:"type,omitempty"`
+}
+
+func (o *Address) GetCity() *string {
+	if o == nil {
+		return nil
+	}
+	return o.City
+}
+
+func (o *Address) GetCountry() *string {
+	if o == nil {
+		return nil
+	}
+	return o.Country
+}
+
+func (o *Address) GetLine1() *string {
+	if o == nil {
+		return nil
+	}
+	return o.Line1
+}
+
+func (o *Address) GetLine2() *string {
+	if o == nil {
+		return nil
+	}
+	return o.Line2
+}
+
+func (o *Address) GetPostalCode() *string {
+	if o == nil {
+		return nil
+	}
+	return o.PostalCode
+}
+
+func (o *Address) GetRegion() *string {
+	if o == nil {
+		return nil
+	}
+	return o.Region
+}
+
+func (o *Address) GetType() *CommerceAddressType {
+	if o == nil {
+		return nil
+	}
+	return o.Type
 }